+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Minimum required PHP version changed to `v8.1`.
 * Adapted CI environment to test PHP `v8.1` and `v8.2`.
+* `ApiResourceServiceProvider` changed to be an aggregate service provider that automatically registers `ETagsServiceProvider`. 
 
 **Non-breaking Changes**